# Use Webpack 5 and enable certain optimizations