## Introduction

Webpack 5 enables better tree-shaking (although the size change is rather small - about 1 kB), certain optimizations and most importantly - **much better subsequent build times**.

## Benchmarks

Cached build time (Webpack 4):

```
npm run build  273,22s user 10,26s system 442% cpu 1:04,13 total
```

Cached build time (Webpack 5):

```
npm run build  42,08s user 6,12s system 223% cpu 21,564 total
```

## Other optimizations

- `reactStrictMode` enforces strict React checks to detect certain issues;
- `experimental.optimizeFonts` optimizes fonts using Webpack 5;
- `experimental.optimizeImages` enables additional image optimization via `next/image` (it currently does nothing, as we are not yet using `next/image`, but will be useful later);
- `experimental.optimizeCss` extracts critical CSS using Webpack 5 and `critters` module;
- `future.strictPostcssConfiguration` enables PostCSS cache to increase subsequent build speed.